Add a scoreboard
Add a HTML element to add winners names into.
Have the server check for a win condition, then emit the winner to all
connected clients.
Have the clients listen for this message (called "reset") that contains
the winners username.
Append this username to the list of winners on the page.