On Click, update the score for all buttons
Now that we add users dynamically, we can't hard code in the button variable
names and switch cases because we can't know what they will be. Or even
how many!

Find all the buttons that match the event name (really should only be one)
and set that buttons score to the event score we have been given.

Then call that buttons update() function to fill the value in the progress
bar.